move specaug function
