+++ espnet/utils/SparseImageWarp.py
@@ -1,4 +1,125 @@
+"""
+SpecAugmnet implementation (https://github.com/zcaceres/spec_augment)
+
+MIT License
+
+Copyright (c) 2019 Zach Caceres
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETjjHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+"""
+
+import random
+
 import torch
+
+
+def time_warp(spec, W=5):
+    """Time warping
+
+    :param torch.Tensor spec: input tensor with shape (T, dim)
+    :param int W int: time warp parameter
+    """
+    spec = spec.unsqueeze(0)
+    num_rows = spec.shape[1]
+    spec_len = spec.shape[2]
+    device = spec.device
+
+    y = num_rows // 2
+    horizontal_line_at_ctr = spec[0][y]
+    assert len(horizontal_line_at_ctr) == spec_len
+
+    point_to_warp = horizontal_line_at_ctr[random.randrange(W, spec_len - W)]
+    assert isinstance(point_to_warp, torch.Tensor)
+
+    # Uniform distribution from (0,W) with chance to be up to W negative
+    dist_to_warp = random.randrange(-W, W)
+    src_pts, dest_pts = (torch.tensor([[[y, point_to_warp]]], device=device),
+                         torch.tensor([[[y, point_to_warp + dist_to_warp]]], device=device))
+    warped_spectro, dense_flows = sparse_image_warp(spec, src_pts, dest_pts)
+    return warped_spectro.squeeze(3).squeeze(0)
+
+
+def freq_mask(spec, F=30, num_masks=1, replace_with_zero=False):
+    """Frequency masking
+
+    :param torch.Tensor spec: input tensor with shape (T, dim)
+    :param int num_masks int: number of masks
+    :param bool replace_with_zero: if True, masked parts will be filled with 0, if False, filled with mean
+    """
+    cloned = spec.unsqueeze(0).clone()
+    num_mel_channels = cloned.shape[1]
+
+    for i in range(0, num_masks):
+        f = random.randrange(0, F)
+        f_zero = random.randrange(0, num_mel_channels - f)
+
+        # avoids randrange error if values are equal and range is empty
+        if (f_zero == f_zero + f):
+            return cloned.squeeze(0)
+
+        mask_end = random.randrange(f_zero, f_zero + f)
+        if (replace_with_zero):
+            cloned[0][f_zero:mask_end] = 0
+        else:
+            cloned[0][f_zero:mask_end] = cloned.mean()
+    return cloned.squeeze(0)
+
+
+def time_mask(spec, T=40, num_masks=1, replace_with_zero=False):
+    """Time masking
+
+    :param torch.Tensor spec: input tensor with shape (T, dim)
+    :param int num_masks int: number of masks
+    :param bool replace_with_zero: if True, masked parts will be filled with 0, if False, filled with mean
+    """
+    cloned = spec.unsqueeze(0).clone()
+    len_spectro = cloned.shape[2]
+
+    for i in range(0, num_masks):
+        t = random.randrange(0, T)
+        t_zero = random.randrange(0, len_spectro - t)
+
+        # avoids randrange error if values are equal and range is empty
+        if (t_zero == t_zero + t):
+            return cloned.squeeze(0)
+
+        mask_end = random.randrange(t_zero, t_zero + t)
+        if (replace_with_zero):
+            cloned[0][:, t_zero:mask_end] = 0
+        else:
+            cloned[0][:, t_zero:mask_end] = cloned.mean()
+    return cloned.squeeze(0)
+
+
+def specaug(spec, num_freq_masks=2, num_time_masks=2):
+    """SpecAugment
+
+    Reference: SpecAugment: A Simple Data Augmentation Method for Automatic Speech Recognition
+        (https://arxiv.org/pdf/1904.08779.pdf)
+
+    This implementation derived from https://github.com/zcaceres/spec_augment
+
+    :param torch.Tensor spec: input tensor with the shape (T, dim)
+    :param int num_freq_masks: number of frequency masks
+    :param int num_time_masks: number of time masks
+    """
+    return time_mask(freq_mask(time_warp(spec), num_masks=num_freq_masks), num_masks=num_time_masks)
 
 
 def sparse_image_warp(img_tensor,
